FEAT: tests pass, still need more testing

+++ services/backend/build/init.sh
@@ -4,6 +4,12 @@ set -eu
 
 mkdir -p /backend/DB/website
 
+if [ "$APP_MODE" = "test" ]; then
+    echo "Running Tests..."
+    exec conda run --no-capture-output -n backend python /backend/manage.py test game.tests
+    exit 0
+fi
+
 rm -f /backend/DB/website/db.sqlite3
 find /backend -path "*/migrations/0*" -delete
 
@@ -13,10 +19,7 @@ conda run -n backend python /backend/manage.py collectstatic --noinput
 
 conda run -n backend bash -c "python /backend/manage.py shell < /backend/seed.py"
 
-if [ "$APP_MODE" = "run" ]; then
-    echo "Starting Production Server..."
-    exec conda run --no-capture-output -n backend daphne -b 0.0.0.0 -p 8000 project.asgi:application
-else
-    echo "Running Tests..."
-    exec conda run --no-capture-output -n backend python /backend/manage.py test game.tests.GameWebsocketTests
-fi
+echo "Starting Production Server..."
+exec conda run --no-capture-output -n backend daphne -b 0.0.0.0 -p 8000 project.asgi:application
+
+exit 0
